requirements.txt
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.30.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.12.0
tqdm>=4.66.0
torchmetrics>=1.3.0
lpips>=0.1.4
